Created Game page for the ParInput component
Created ParInput component to record player data for each hole on the course selected. Scores are saved as an array and stored in scores state. Successful POST request for adding the finished game and routes to the recap page

Added /recap Route in App.js

Added styling to components found on the Game page